Fixing the inclusion order of the extensions tasks.
If you add extensions to your postgres config and run the roles as
it was, it will fail to bring up the daemon, because the task that
actually installs the extensions is ran only after the task that
starts the daemon. As such, brought up the exclusions.yml so that
it runs before the daemon start.